Convert to FP32

+++ algorithms/triad/triadAlgorithm.h
@@ -7,15 +7,15 @@
 
 class TriadAlgorithm final {
    public:
-    Eigen::Vector3d update(const Eigen::Vector3d& rHat_SB_N,
-                           const Eigen::Vector3d& hReqHat_N,
-                           const Eigen::Vector3d& hRefHat_B) const;
+    Eigen::Vector3f update(const Eigen::Vector3f& rHat_SB_N,
+                           const Eigen::Vector3f& hReqHat_N,
+                           const Eigen::Vector3f& hRefHat_B) const;
 
-    void setA1Hat_B(const Eigen::Vector3d& a1Hat_B);
-    Eigen::Vector3d getA1Hat_B() const;
+    void setA1Hat_B(const Eigen::Vector3f& a1Hat_B);
+    Eigen::Vector3f getA1Hat_B() const;
 
    private:
-    Eigen::Vector3d a1Hat_B = Eigen::Vector3d::Zero();
+    Eigen::Vector3f a1Hat_B = Eigen::Vector3f::Zero();
 };
 
 #endif
